Move remembered approval rules to core
